Server: restructure directories
